+++ tools/testoperator/src/tests/mod.rs
@@ -14,32 +14,33 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use std::collections::BTreeMap;
+use std::{collections::BTreeMap, path::Path};
 
-use crate::commands::DatasetTestArgs;
+use crate::commands::{CommonArgs, DatasetTestArgs};
 use test_framework::{
     anyhow, app::App, spiced::StartRequest, spicepod::Spicepod, spicepod_utils::from_app,
 };
 
 pub(crate) mod bench;
 pub(crate) mod data_consistency;
-pub(crate) mod http_consistency;
+pub(crate) mod http;
 pub(crate) mod load;
 pub(crate) mod throughput;
 
 pub(crate) type RowCounts = BTreeMap<String, usize>;
 
 pub(crate) fn get_app_and_start_request(
-    args: &DatasetTestArgs,
+    args: &CommonArgs,
+    data_dir: Option<&Path>,
 ) -> anyhow::Result<(App, StartRequest)> {
-    let spicepod = Spicepod::load_exact(args.common.spicepod_path.clone())?;
+    let spicepod = Spicepod::load_exact(args.spicepod_path.clone())?;
     let app = test_framework::app::AppBuilder::new(spicepod.name.clone())
         .with_spicepod(spicepod)
         .build();
 
-    let start_request = StartRequest::new(args.common.spiced_path.clone(), from_app(app.clone()))?;
-    let start_request = if let Some(data_dir) = &args.data_dir {
-        start_request.with_data_dir(data_dir.clone())
+    let start_request = StartRequest::new(args.spiced_path.clone(), from_app(app.clone()))?;
+    let start_request = if let Some(data_dir) = data_dir {
+        start_request.with_data_dir(data_dir.to_path_buf())
     } else {
         start_request
     };
@@ -48,7 +49,7 @@ pub(crate) fn get_app_and_start_request(
 }
 
 pub(crate) fn env_export(args: &DatasetTestArgs) -> anyhow::Result<()> {
-    let (_, mut start_request) = get_app_and_start_request(args)?;
+    let (_, mut start_request) = get_app_and_start_request(&args.common, args.data_dir.as_deref())?;
 
     start_request.prepare()?;
     let tempdir_path = start_request.get_tempdir_path();
